+++ backend/canisters/notifications/impl/src/updates/c2c_push_notifications.rs
@@ -6,8 +6,8 @@ use notifications_canister::c2c_push_notifications::{Response::*, *};
 use serde_bytes::ByteBuf;
 use std::collections::HashSet;
 use types::{
-    BotNotification, BotNotificationEnvelope, CanPushNotificationsArgs, CanPushNotificationsResponse, CanisterId,
-    IdempotentEnvelope, Notification, NotificationEnvelope, TimestampMillis, UserId, UserNotificationEnvelope,
+    BotEventWrapper, BotNotification, BotNotificationEnvelope, CanPushNotificationsArgs, CanPushNotificationsResponse,
+    CanisterId, IdempotentEnvelope, Notification, NotificationEnvelope, TimestampMillis, UserId, UserNotificationEnvelope,
 };
 
 #[update(msgpack = true)]
@@ -66,7 +66,7 @@ fn c2c_push_notifications_impl(
                 Notification::User(n) => {
                     push_user_notification(n.sender, n.recipients, n.notification_bytes, now, &mut state.data);
                 }
-                Notification::Bot(n) => push_bot_notification(n, now, &mut state.data),
+                Notification::Bot(n) => push_bot_notification(n, state.env.canister_id(), now, &mut state.data),
             }
         }
     }
@@ -96,12 +96,20 @@ pub(crate) fn push_user_notification(
     }
 }
 
-fn push_bot_notification(mut notification: BotNotification, now: TimestampMillis, data: &mut Data) {
+fn push_bot_notification(
+    mut notification: BotNotification,
+    this_canister_id: CanisterId,
+    now: TimestampMillis,
+    data: &mut Data,
+) {
     notification.recipients.retain(|r| data.bot_endpoints.contains_key(r));
 
     if !notification.recipients.is_empty() {
         data.notifications.add(NotificationEnvelope::Bot(BotNotificationEnvelope {
-            event: notification.event,
+            event: BotEventWrapper {
+                api_gateway: this_canister_id,
+                event: notification.event,
+            },
             recipients: notification.recipients,
             timestamp: now,
         }));