+++ backend/libraries/types/src/notifications.rs
@@ -45,6 +45,12 @@ pub enum BotEvent {
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct BotEventWrapper {
+    pub api_gateway: CanisterId,
+    pub event: BotEvent,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct BotChatEvent {
     #[serde(rename = "e")]
     pub event_type: ChatEventType,
@@ -72,22 +78,10 @@ pub struct BotCommunityEvent {
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub enum BotLifecycleEvent {
-    #[serde(rename = "r")]
-    Registered(BotRegisteredEvent),
-    #[serde(rename = "d")]
-    Removed,
     #[serde(rename = "i")]
     Installed(BotInstalledEvent),
     #[serde(rename = "u")]
     Uninstalled(BotUninstalledEvent),
-}
-
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-pub struct BotRegisteredEvent {
-    #[serde(rename = "i")]
-    pub bot_id: UserId,
-    #[serde(rename = "n")]
-    pub bot_name: String,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
@@ -141,7 +135,7 @@ pub struct UserNotificationEnvelope {
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct BotNotificationEnvelope {
     #[serde(rename = "e")]
-    pub event: BotEvent,
+    pub event: BotEventWrapper,
     #[serde(rename = "r")]
     pub recipients: Vec<UserId>,
     #[serde(rename = "t")]